Add Calendar Interface and Upcoming Function
Adds the Calendar interface and a Register function to allow a
unified query of calendar sources. Adds an Upcoming function to
merge the results of the Upcoming method on all registered Calendars.

The Upcoming function assumes that the results of Upcoming method
calls are sorted according to the desc argument, allowing a simple
merge instead of a fresh sort.